requirements.txt
pytest
langchain-ollama
langchain-community
python-dotenv
requests
yfinance
pipdeptree
langchain
google-search-results



